+++ matcher/exact.go
@@ -1,23 +1,23 @@
 package matcher
 
 import (
-	gn "github.com/gnames/gnames/domain/entity"
-	"github.com/gnames/gnmatcher/domain/entity"
+	mlib "github.com/gnames/gnlib/domain/entity/matcher"
+	vlib "github.com/gnames/gnlib/domain/entity/verifier"
 )
 
 // Match tries to match a canonical form of a name-string exactly to canonical
 // from from gnames database.
-func (m Matcher) Match(ns NameString) *entity.Match {
+func (m Matcher) Match(ns NameString) *mlib.Match {
 	var isIn bool
 	m.Filters.Mux.Lock()
 	isIn = m.Filters.Canonical.Check([]byte(ns.CanonicalID))
 	m.Filters.Mux.Unlock()
 	if isIn {
-		return &entity.Match{
+		return &mlib.Match{
 			ID:        ns.ID,
 			Name:      ns.Name,
-			MatchType: gn.Exact,
-			MatchItems: []entity.MatchItem{
+			MatchType: vlib.Exact,
+			MatchItems: []mlib.MatchItem{
 				{
 					ID:       ns.CanonicalID,
 					MatchStr: ns.Canonical,
@@ -30,18 +30,18 @@ func (m Matcher) Match(ns NameString) *entity.Match {
 
 // MatchVirus tries to match a name-string exactly to a virus name from the
 // gnames database.
-func (m Matcher) MatchVirus(ns NameString) *entity.Match {
+func (m Matcher) MatchVirus(ns NameString) *mlib.Match {
 	var isIn bool
 	m.Filters.Mux.Lock()
 	isIn = m.Filters.Virus.Check([]byte(ns.ID))
 	m.Filters.Mux.Unlock()
 	if isIn {
-		return &entity.Match{
+		return &mlib.Match{
 			ID:         ns.ID,
 			Name:       ns.Name,
 			VirusMatch: true,
-			MatchType:  gn.Exact,
-			MatchItems: []entity.MatchItem{
+			MatchType:  vlib.Exact,
+			MatchItems: []mlib.MatchItem{
 				{
 					ID:       ns.ID,
 					MatchStr: ns.Name,