+++ matcher/matcher.go
@@ -5,13 +5,12 @@ import (
 	"sync"
 
 	"github.com/dgraph-io/badger/v2"
-	gn "github.com/gnames/gnames/domain/entity"
+	mlib "github.com/gnames/gnlib/domain/entity/matcher"
+	vlib "github.com/gnames/gnlib/domain/entity/verifier"
 	"github.com/gnames/gnlib/sys"
 	"github.com/gnames/gnmatcher/bloom"
 	"github.com/gnames/gnmatcher/config"
 	"github.com/gnames/gnmatcher/dbase"
-	"github.com/gnames/gnmatcher/domain/entity"
-	"github.com/gnames/gnmatcher/domain/usecase"
 	"github.com/gnames/gnmatcher/fuzzy"
 	"github.com/gnames/gnmatcher/stemskv"
 	uuid "github.com/satori/go.uuid"
@@ -24,7 +23,7 @@ var (
 	// GNUUID is a UUID seed made from 'globalnames.org' domain to generate
 	// UUIDv5 identifiers.
 	GNUUID    = uuid.NewV5(uuid.NamespaceDNS, "globalnames.org")
-	nilResult *entity.Match
+	nilResult *mlib.Match
 )
 
 // Matcher contains data and functions necessary for exact, fuzzy and partial
@@ -33,7 +32,7 @@ type Matcher struct {
 	Config  config.Config
 	Filters *bloom.Filters
 	KeyVal  *badger.DB
-	usecase.FuzzyMatcher
+	mlib.FuzzyMatcher
 }
 
 // MatchTask contains a name to be matched and an index where it should be
@@ -45,7 +44,7 @@ type MatchTask struct {
 
 type MatchResult struct {
 	Index int
-	Match *entity.Match
+	Match *mlib.Match
 }
 
 // NewMatcher creates a new instance of Matcher struct.
@@ -81,7 +80,7 @@ func (m Matcher) MatchWorker(chIn <-chan MatchTask,
 	chOut chan<- MatchResult, wg *sync.WaitGroup) {
 	parser := gnparser.NewGNparser()
 	defer wg.Done()
-	var matchResult *entity.Match
+	var matchResult *mlib.Match
 
 	for tsk := range chIn {
 		ns, parsed := NewNameString(parser, tsk.Name)
@@ -107,16 +106,16 @@ func (m Matcher) MatchWorker(chIn <-chan MatchTask,
 // DetectAbbreviated checks if parsed name is abbreviated. If name is not
 // abbreviated the function returns nil. If it is abbreviated, it returns
 // result with the MatchType 'NONE'.
-func DetectAbbreviated(parsed *pb.Parsed) *entity.Match {
+func DetectAbbreviated(parsed *pb.Parsed) *mlib.Match {
 	if parsed.Quality != int32(3) {
 		return nilResult
 	}
 	for _, v := range parsed.QualityWarning {
 		if strings.HasPrefix(v.Message, "Abbreviated") {
-			return &entity.Match{
+			return &mlib.Match{
 				ID:        parsed.Id,
 				Name:      parsed.Verbatim,
-				MatchType: gn.NoMatch,
+				MatchType: vlib.NoMatch,
 			}
 		}
 	}
@@ -134,10 +133,10 @@ func (m Matcher) prepareWorkDirs() {
 	}
 }
 
-func emptyResult(ns NameString) *entity.Match {
-	return &entity.Match{
+func emptyResult(ns NameString) *mlib.Match {
+	return &mlib.Match{
 		ID:        ns.ID,
 		Name:      ns.Name,
-		MatchType: gn.NoMatch,
+		MatchType: vlib.NoMatch,
 	}
 }